Use tox to test
