+++ .github/workflows/pythonpackage.yml
@@ -18,7 +18,8 @@ jobs:
         pre-commit install --install-hooks --config .pre-commit-config-ci.yaml
     - run: pre-commit run --all-files --config .pre-commit-config-ci.yaml
 
-  tests:
+  check-pipfile:
+    name: Check Pipfile.lock is up-to-date
     runs-on: ubuntu-latest
     strategy:
       matrix:
@@ -34,5 +35,40 @@ jobs:
         python -m pip install --upgrade pip
         pip install pipenv
         pipenv --python ${{ matrix.python-version }}
-        pipenv install --deploy --dev
-    - run: pipenv run make test
+    - run: pipenv install --deploy --dev
+
+  tox:
+    name: Unit tests
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [3.6, 3.7]
+    steps:
+    - uses: actions/checkout@v1
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v1
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Set up Pipenv
+      run: |
+        python -m pip install --upgrade pip
+        pip install tox
+    - run: tox -e ${{ matrix.python-version }}
+
+  tox-e2e:
+    name: E2E tests
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [3.6, 3.7]
+    steps:
+    - uses: actions/checkout@v1
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v1
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Set up Pipenv
+      run: |
+        python -m pip install --upgrade pip
+        pip install tox
+    - run: tox -e e2e
