+++ tox.ini
@@ -1,0 +1,24 @@
+[tox]
+envlist = 3.6,3.7
+
+[testenv]
+basepython =
+    3.6: python3.6
+    3.7: python3.7
+deps = pytest
+commands =
+    pytest tests
+
+[testenv:e2e]
+basepython = python
+whitelist_externals=
+    docker
+deps = pytest
+commands_pre =
+    docker pull kodebox/codechain:ebacfe7b74296cc1e793cf3238748b7280f8072a
+    docker run --name tox-e2e --rm -d -p 8080:8080 kodebox/codechain:ebacfe7b74296cc1e793cf3238748b7280f8072a --jsonrpc-interface 0.0.0.0 -c solo --reseal-min-period 0
+    docker ps -a
+commands =
+    pytest tests-e2e
+commands_post =
+    docker stop tox-e2e
